refactor(settings): default Content() for settings + dev-settings + featureflag blocs
Move SettingsScreen, AppSettingsScreen, DeveloperSettingsScreen, and
FeatureFlagsScreen into their respective `public` modules, and give the
five Bloc interfaces (Settings, AppSettings, SettingsRoot, DeveloperSettings,
FeatureFlags) a default `Content()` that delegates to the screen — so the
impl Blocs no longer override `Content` themselves. Inline preview fakes
and `AppSettingsPreviews` follow suit.
