+++ client/developer-settings/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/devsettings/DeveloperSettingsBloc.kt
@@ -1,8 +1,11 @@
 package com.plusmobileapps.chefmate.devsettings
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.Environment
+import com.plusmobileapps.chefmate.devsettings.ui.DeveloperSettingsScreen
 import com.plusmobileapps.chefmate.ui.ComposeScreen
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
@@ -10,6 +13,11 @@ import kotlinx.coroutines.flow.StateFlow
 
 interface DeveloperSettingsBloc : ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        DeveloperSettingsScreen(bloc = this, modifier = modifier)
+    }
 
     fun onBack()
