+++ client/settings/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/settings/impl/AppSettingsBlocImpl.kt
@@ -1,14 +1,11 @@
 package com.plusmobileapps.chefmate.settings.impl
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.di.AppScope
 import com.plusmobileapps.chefmate.getViewModel
 import com.plusmobileapps.chefmate.settings.AppSettingsBloc
 import com.plusmobileapps.chefmate.settings.AppSettingsBloc.Output
-import com.plusmobileapps.chefmate.settings.impl.ui.AppSettingsScreen
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
 import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedInject
@@ -64,10 +61,5 @@ class AppSettingsBlocImpl(
 
     override fun onRecipeCategoriesClicked() {
         output.onNext(Output.OpenRecipeCategories)
-    }
-
-    @Composable
-    override fun Content(modifier: Modifier) {
-        AppSettingsScreen(bloc = this, modifier = modifier)
     }
 }
