+++ client/settings/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/settings/impl/ui/AppSettingsPreviews.kt
@@ -1,11 +1,11 @@
 package com.plusmobileapps.chefmate.settings.impl.ui
 
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.arkivanov.essenty.backhandler.BackDispatcher
 import com.arkivanov.essenty.backhandler.BackHandler
 import com.plusmobileapps.chefmate.settings.AppSettingsBloc
+import com.plusmobileapps.chefmate.settings.ui.AppSettingsScreen
 import com.plusmobileapps.chefmate.ui.theme.ChefMateTheme
 import kotlinx.coroutines.flow.MutableStateFlow
 
@@ -31,11 +31,6 @@ private fun appSettingsBloc(model: AppSettingsBloc.Model): AppSettingsBloc =
         override fun onExportRecipesClicked() = Unit
 
         override fun onRecipeCategoriesClicked() = Unit
-
-        @Composable
-        override fun Content(modifier: Modifier) {
-            AppSettingsScreen(bloc = this, modifier = modifier)
-        }
     }
 
 val previewAppSettingsBloc: AppSettingsBloc =
