+++ client/settings/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/settings/AppSettingsBloc.kt
@@ -1,13 +1,21 @@
 package com.plusmobileapps.chefmate.settings
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.arkivanov.essenty.backhandler.BackHandlerOwner
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
+import com.plusmobileapps.chefmate.settings.ui.AppSettingsScreen
 import com.plusmobileapps.chefmate.ui.ComposeScreen
 import kotlinx.coroutines.flow.StateFlow
 
 interface AppSettingsBloc : BackHandlerOwner, ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        AppSettingsScreen(bloc = this, modifier = modifier)
+    }
 
     fun onBack()
