+++ client/settings/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/settings/ui/AppSettingsScreen.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.chefmate.settings.impl.ui
+package com.plusmobileapps.chefmate.settings.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
