+++ client/settings/root/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/settings/root/SettingsRootBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.settings.root
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.value.Value
 import com.arkivanov.essenty.backhandler.BackHandlerOwner
@@ -15,6 +17,11 @@ import com.plusmobileapps.chefmate.ui.ComposeScreen
 
 interface SettingsRootBloc : BackHandlerOwner, BackClickBloc, ComposeScreen {
     val routerState: Value<ChildStack<*, Child>>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        SettingsRootScreen(bloc = this, modifier = modifier)
+    }
 
     sealed class Child {
